; cperl-mode.el: Fix indentation issues after a format declaration.
* lisp/progmodes/cperl-mode.el (cperl-sniff-for-indent): Recognize
the end of a format declaration as a statement boundary.  This fixes
indentation if several format declarations immediately follow each
other.
(cperl-find-pods-heres): use the correct capture group for the
name of an unterminated format declaration.  Fix an out of
boundary error when a format declaration is typed at the end of a
buffer (Bug#66139).